third_party/blink/web_tests/external/wpt/workers/interfaces/SharedWorkerGlobalScope/onconnect-expected.txt

This is a testharness.js-based test.
[FAIL] onconnect
  assert_array_equals: expected property 1 to be "null" but got "[object Object]" (expected array ["null", "null", "function", ""] got ["null", "[object Object]", "function", null])
Harness: the test ran to completion.
